refactor(skills): split package into focused files
skills.go (402 -> 82 lines) now only holds the public API: the Skill type

and the Load entry point. The implementation is split into focused files:

- local.go: filesystem skill discovery, with a declarative localSearchPaths

  table that makes the precedence rules (codex, claude, agents, project) easy

  to read at a glance. findGitRoot+projectSearchDirs share a single walk-up.

- frontmatter.go: YAML-ish frontmatter parsing, split into parseTopLevelLine

  and parseIndentedLine for clarity.

Other small simplifications:

- Load now sorts its result by name for deterministic output.

- Load avoids the double map -> slice -> map round-trip via loadLocalSkillsInto.

- Skill.IsFork uses a value receiver for consistency with the rest of the API.

- isHTTPSource extracted from Load.

No public API or behavior change; all existing tests pass.
